Remove PasswordCreation component and generate 3 unique passwords
