--- client/src/PageLayout/Components/PasswordWalkthrough/Components/PasswordCreation/index.ts
@@ -1,3 +1,0 @@
-import PasswordCreation from "./PasswordCreation";
-
-export default PasswordCreation;
